fix[41]: fix server

+++ server/models/Answer.ts
@@ -27,18 +27,18 @@ class Answer
 Answer.init(
   {
     id: {
-      type: DataTypes.INTEGER.UNSIGNED,
+      type: DataTypes.BIGINT.UNSIGNED,
       autoIncrement: true,
       primaryKey: true,
     },
     responseId: {
-      type: DataTypes.INTEGER.UNSIGNED,
+      type: DataTypes.BIGINT.UNSIGNED,
       allowNull: false,
       references: { model: Response, key: 'id' },
       field: 'response_id',
     },
     questionId: {
-      type: DataTypes.INTEGER.UNSIGNED,
+      type: DataTypes.BIGINT.UNSIGNED,
       allowNull: false,
       references: { model: Question, key: 'id' },
       field: 'question_id',
