+++ server/models/Question.ts
@@ -37,7 +37,7 @@ class Question
 Question.init(
   {
     id: {
-      type: DataTypes.INTEGER.UNSIGNED,
+      type: DataTypes.BIGINT.UNSIGNED,
       autoIncrement: true,
       primaryKey: true,
     },
@@ -71,7 +71,7 @@ Question.init(
       defaultValue: true,
     },
     templateId: {
-      type: DataTypes.INTEGER.UNSIGNED,
+      type: DataTypes.BIGINT.UNSIGNED,
       allowNull: false,
       references: {
         model: Template,
